SpokenLanguageIdentification/requirements.txt
loguru
librosa
numpy
torch
torchvision
pyyaml